activate snapshot handlers
